% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{read_session_xml}
\alias{read_session_xml}
\title{Read Session XML}
\usage{
read_session_xml(xml_file_name)
}
\arguments{
\item{xml_file_name}{A string. The name of the XML file to be read.}
}
\value{
A dataframe with columns: 'session', 'eventCategory', 'eventAction', 'eventName', 'timestamp', 'serverTimePretty'.
}
\description{
This function reads an XML file and extracts session recording details.
It returns a dataframe with session details including event category, action, name, timestamp, and server time.
}
\examples{
\dontrun{
read_session_xml("session.xml")
}

}
